[chore] Address deprecation of JavaExecSpec.main property
Gradle has deprecated the `JavaExecSpec.getMain()` and `JavaExecSpec.setMain()`
methods in favour of the new `mainClass` property.
The old methods are scheduled to be removed in Gradle 8.0, but Gradle already
warns about their deprecation when used.

This change addresses this issue by using the `mainClass` property if it
exists and fallback to the `setMain` when it doesn't to provide
compatibility with older Gradle versions.