Allow building reproducible Jandex indexes
Reproducible builds are becoming more and more important to enable the validation of all built artifacts. TL;DR all built artifacts are be 100% equal "byte-by-byte".
This also requires that the serialized Jandex indexes to be reproducible.

Since Jandex version 3.5.0, its `IndexWriter` builds reproducible serialized indexes, but only if the `.class` files are indexed (`Indexer.index()`) in a deterministic order.

This change updates the `JandexWorkAction` to first collect all `.class` files to be indexed in a platform agnostic way and then lets Jandex index the `.class` files in a deterministic order.
`Path.toURI()` is used as the "sort key", as this representation is file-separator agnostic (`/` vs `\`) and it is (should be?) file-system agnostic.